:sparkles: add clerk
