+++ package.json
@@ -9,6 +9,8 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@clerk/localizations": "^1.26.16",
+    "@clerk/nextjs": "^5.0.0-beta.34",
     "firebase": "^10.8.0",
     "next": "14.1.0",
     "react": "^18",
